Create HDS_background_many_potentials
Remake of the original HDS code for the background. Potentials are now objects that can be declared with an arbitrary number of parameters. The shooting method remains the secant, but a fallback for the bisection method is implemented when it fails.
All the rest is working but the bisection, I am still trying to make it right, so far I don't know why it is failling.